refactor: migrate DialogUtils to ProgressBarDialogIndeterminate
